requirements.txt
Flask>=2.3.3
Werkzeug>=2.3.7
Jinja2>=3.1.2
MarkupSafe>=2.1.3
itsdangerous>=2.1.2
click>=8.1.7
blinker>=1.6.3
PyJWT>=2.8.0
redis>=5.0.1
psutil>=5.9.6
Flask-WTF>=1.1.1
pyotp>=2.9.0
requests>=2.31.0
bcrypt>=4.0.1
python-dotenv>=1.0.0 